+++ ReleaseNotes.md
@@ -11,6 +11,7 @@ Version 0.1.12
 * Add minimal page blob abstraction so that lists don't blow up when the container includes a page blob.
 * Add flag that causes LightBlue hosts to provide the hosted storage implementation (via the LightBlue APIs)
 * Add additional async variations of storage API methods for which async support makes sense.
+* Listing of blobs will load metadata when the BlobListingDetails.Metadata flag is specified for standalone containers and directories.
 
 ###Implementation
 * Removed the StandaloneAzureBlockBlob(Uri) constructor.